refactor: move logic into runtime
